Formulaire Produit OK Formulaire Categorie OK
T.A.F
Formulaire Client

+++ database/migrations/2017_03_02_131038_vente.php
@@ -15,7 +15,7 @@ class Vente extends Migration
         //
         Schema::create('ventes', function (Blueprint $table) {
             $table->increments('id');
-            $table->double('quantite');
+            $table->integer('quantite');
             $table->date('date');
             $table->string('type');
             $table->integer('visite_id')->unsigned();
